requirements.txt
appdirs==1.4.3
BTrees==4.4.1
Django==1.11
django-filter==1.0.2
djangorestframework==3.6.2
drink==0.0.10
enum34==1.1.6
fs==2.0.3
futures==3.1.1
httpie==0.9.9
Jinja2==2.9.6
Markdown==2.6.8
MarkupSafe==1.0
packaging==16.8
persistent==4.2.4.2
Pygments==2.2.0
pyparsing==2.2.0
pytz==2017.2
requests==2.13.0
scandir==1.5
six==1.10.0
transaction==2.1.2
trollius==2.1
Whoosh==2.7.4
zc.lockfile==1.2.1
ZConfig==3.1.0
zdaemon==4.2.0
ZEO==5.1.0
ZODB==5.2.3
ZODB3==3.11.0
zodbpickle==0.6.0
zope.interface==4.4.0
